+++ axum/src/routing/tests/mod.rs
@@ -3,12 +3,12 @@ use crate::{
     extract::{self, Path},
     handler::Handler,
     response::IntoResponse,
-    routing::{any, delete, get, on, patch, post, service_method_routing as service, MethodFilter},
+    routing::{delete, get, get_service, on, on_service, patch, patch_service, post, MethodFilter},
     test_helpers::*,
     BoxError, Json, Router,
 };
 use bytes::Bytes;
-use http::{header::HeaderMap, Method, Request, Response, StatusCode, Uri};
+use http::{Method, Request, Response, StatusCode, Uri};
 use hyper::Body;
 use serde::Deserialize;
 use serde_json::{json, Value};
@@ -135,20 +135,20 @@ async fn routing_between_services() {
     let app = Router::new()
         .route(
             "/one",
-            service::get(service_fn(|_: Request<Body>| async {
+            get_service(service_fn(|_: Request<Body>| async {
                 Ok::<_, Infallible>(Response::new(Body::from("one get")))
             }))
-            .post(service_fn(|_: Request<Body>| async {
+            .post_service(service_fn(|_: Request<Body>| async {
                 Ok::<_, Infallible>(Response::new(Body::from("one post")))
             }))
-            .on(
+            .on_service(
                 MethodFilter::PUT,
                 service_fn(|_: Request<Body>| async {
                     Ok::<_, Infallible>(Response::new(Body::from("one put")))
                 }),
             ),
         )
-        .route("/two", service::on(MethodFilter::GET, any(handle)));
+        .route("/two", on_service(MethodFilter::GET, handle.into_service()));
 
     let client = TestClient::new(app);
 
@@ -202,7 +202,7 @@ async fn service_in_bottom() {
         Ok(Response::new(hyper::Body::empty()))
     }
 
-    let app = Router::new().route("/", service::get(service_fn(handler)));
+    let app = Router::new().route("/", get_service(service_fn(handler)));
 
     TestClient::new(app);
 }
@@ -248,8 +248,8 @@ async fn wrong_method_service() {
     }
 
     let app = Router::new()
-        .route("/", service::get(Svc).post(Svc))
-        .route("/foo", service::patch(Svc));
+        .route("/", get_service(Svc).post_service(Svc))
+        .route("/foo", patch_service(Svc));
 
     let client = TestClient::new(app);
 
@@ -507,17 +507,6 @@ async fn route_layer() {
 
 #[tokio::test]
 #[should_panic(
-    expected = "Invalid route: insertion failed due to conflict with previously registered route: /foo"
-)]
-async fn conflicting_route() {
-    let app = Router::new()
-        .route("/foo", get(|| async {}))
-        .route("/foo", get(|| async {}));
-    TestClient::new(app);
-}
-
-#[tokio::test]
-#[should_panic(
     expected = "Invalid route: insertion failed due to conflict with previously registered route: /*axum_nest. Note that `nest(\"/\", _)` conflicts with all routes. Use `Router::fallback` instead"
 )]
 async fn good_error_message_if_using_nest_root() {
@@ -536,4 +525,44 @@ async fn good_error_message_if_using_nest_root_when_merging() {
     let two = Router::new().route("/", get(|| async {}));
     let app = one.merge(two);
     TestClient::new(app);
+}
+
+#[tokio::test]
+async fn different_methods_added_in_different_routes() {
+    let app = Router::new()
+        .route("/", get(|| async { "GET" }))
+        .route("/", post(|| async { "POST" }));
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/").send().await;
+    let body = res.text().await;
+    assert_eq!(body, "GET");
+
+    let res = client.post("/").send().await;
+    let body = res.text().await;
+    assert_eq!(body, "POST");
+}
+
+#[tokio::test]
+async fn different_methods_added_in_different_routes_deeply_nested() {
+    let app = Router::new()
+        .route("/foo/bar/baz", get(|| async { "GET" }))
+        .nest(
+            "/foo",
+            Router::new().nest(
+                "/bar",
+                Router::new().route("/baz", post(|| async { "POST" })),
+            ),
+        );
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar/baz").send().await;
+    let body = res.text().await;
+    assert_eq!(body, "GET");
+
+    let res = client.post("/foo/bar/baz").send().await;
+    let body = res.text().await;
+    assert_eq!(body, "POST");
 }