+++ examples/websockets/src/main.rs
@@ -7,14 +7,13 @@
 //! ```
 
 use axum::{
-    error_handling::HandleErrorExt,
     extract::{
         ws::{Message, WebSocket, WebSocketUpgrade},
         TypedHeader,
     },
     http::StatusCode,
     response::IntoResponse,
-    routing::{get, service_method_routing as service},
+    routing::{get, get_service},
     Router,
 };
 use std::net::SocketAddr;
@@ -34,7 +33,7 @@ async fn main() {
     // build our application with some routes
     let app = Router::new()
         .fallback(
-            service::get(
+            get_service(
                 ServeDir::new("examples/websockets/assets").append_index_html_on_directories(true),
             )
             .handle_error(|error: std::io::Error| {